reoptimized report generation code
