+++ APTRS/project/report.py
@@ -1,6 +1,5 @@
 import io
 import logging
-import ssl
 import urllib
 import os
 import requests
@@ -13,7 +12,7 @@ from django.http import HttpResponse
 from django.template.loader import render_to_string
 from django.template import TemplateDoesNotExist
 from pygal.style import Style
-from weasyprint import HTML, default_url_fetcher
+from weasyprint import HTML
 from xlsxwriter.workbook import Workbook
 from docx import Document
 from docxtpl import DocxTemplate,RichText, InlineImage
@@ -29,104 +28,33 @@ from accounts.models import CustomUser, CustomGroup
 from customers.models import Company
 from .models import (PrjectScope, Project, ProjectRetest, Vulnerability,
                      Vulnerableinstance)
-from utils.image_parsing import fetch_image_bytes, find_images
+from utils.image_parsing import find_images
+from utils.doc_style import get_subdoc ,main_doc_style, apply_font_to_elements , resize_inline_images
 logger = logging.getLogger(__name__)
 logger = logging.getLogger('weasyprint')
 
+# Global variable to store base_url and token
+base_url = ""
 token = None
 
-def apply_font_style(element, font_name, font_size):
-    if hasattr(element, 'font'):
-        element.font.name = font_name
-        element.font.size = Pt(font_size)
 
-def apply_font_to_elements(elements, font_name, font_size):
-    for element in elements:
-        apply_font_style(element, font_name, font_size)
-        if hasattr(element, 'runs'):
-            for run in element.runs:
-                apply_font_style(run, font_name, font_size)
+def CheckReport(Report_format,Report_type,pk,url,standard,request):
+    global base_url
+    base_url = url
+    if Report_format == "excel":
+        response =  CreateExcel(pk)
 
-def resize_inline_images(temp_doc, fixed_width):
-    for i, image in enumerate(temp_doc.inline_shapes):
-        original_width, original_height = image.width, image.height
+    global token
+    auth_header = request.headers.get('Authorization')
+    if auth_header:
+        token = auth_header.split(' ')[1] if auth_header.startswith('Bearer ') else None
+    if Report_format == "docx":
+        response = generate_vulnerability_document(pk,Report_type,standard)
+    if Report_format == "pdf": 
+        response = GetHTML(Report_type,pk,standard,request)
+    
+    return response
 
-        # Calculate new height to maintain the aspect ratio
-        new_height = int(original_height * fixed_width / original_width)
-
-        # Set the fixed width and calculated height
-        image.width = fixed_width
-        image.height = new_height
-
-def get_subdoc(doc, raw_html):
-
-    temp_doc = Document()
-    temp_parser = HtmlToDocx()
-
-    if raw_html is not None:
-
-        images = find_images(raw_html, settings.USE_S3, base_url)
-
-        # Convert image src paths - doctpl does not support loading img over url, adding image full path
-        for i, img in enumerate(images):
-            raw_html = raw_html.replace(img['html_tag'], f"{{{{img{i}}}}}")
-
-        # Convert HTML to temporary DOCX
-
-
-        temp_parser.add_html_to_document(raw_html, temp_doc)
-
-        # Resize images in the temporary DOCX
-        ## https://stackoverflow.com/questions/76571366/resizing-all-images-in-a-word-document-using-python
-        text_width = temp_doc.sections[0].page_width - temp_doc.sections[0].left_margin - temp_doc.sections[0].right_margin
-
-        resize_inline_images(temp_doc, fixed_width=text_width)
-        apply_font_to_elements(temp_doc.element.body, 'Calibri', 16)
-        
-        for paragraph in temp_doc.paragraphs:
-            paragraph.paragraph_format.space_before = Pt(5)
-            paragraph.paragraph_format.space_after = Pt(5)
-            paragraph.paragraph_format.left_indent = Pt(45)
-            paragraph.paragraph_format.right_indent = Inches(1)
-            paragraph.paragraph_format.top_indent = Inches(1)
-            paragraph.paragraph_format.line_spacing = 1.5
-
-        
-        obj_styles = temp_doc.styles
-        for current_style in obj_styles:
-            element = current_style
-            if hasattr(element, 'font'):
-                font = element.font
-                font.name = 'Calibri'
-                font.size = Pt(16)
-
-
-        font = temp_doc.styles['Normal'].font
-        font.name = 'Calibri'
-        font.size = Pt(16)
-        font = temp_doc.styles['List Bullet'].font
-        font.name = 'Calibri'
-        font.size = Pt(16)
-        
-
-        # Save temporary DOCX in memory
-        subdoc_tmp = io.BytesIO()
-        temp_doc.save(subdoc_tmp)
-        subdoc_tmp.seek(0)
-        sub_docxtpl = DocxTemplate(subdoc_tmp)
-        context = {}
-        for i, img in enumerate(images):
-            img_obj = InlineImage(sub_docxtpl, img['bytes'])
-            context[f"img{i}"] = img_obj
-        
-        sub_docxtpl.render(context)
-        sub_docxtpl.save(subdoc_tmp)
-        subdoc_tmp.seek(0)
-
-
-        # Create docxtpl subdoc object
-        subdoc = doc.new_subdoc(subdoc_tmp)
-        return subdoc
 
 
 def generate_vulnerability_document(pk,Report_type,standard):
@@ -145,15 +73,15 @@ def generate_vulnerability_document(pk,Report_type,standard):
     projectmanagers = CustomUser.objects.filter(groups=project_manager_group)
     customeruser =  CustomUser.objects.filter(company=project.companyname, is_active=True)
     mycomany = Company.objects.filter(internal=True).values_list('name', flat=True).first()
-    project_description = get_subdoc(doc, project.description)
-    project_exception = get_subdoc(doc, project.projectexception)
+    project_description = get_subdoc(doc, project.description,token, base_url)
+    project_exception = get_subdoc(doc, project.projectexception,token, base_url)
 
     for vulnerability in vuln:
         # Convert CKEditor fields from HTML to DOCX format
-        vulnerability.vulnerabilitydescription = get_subdoc(doc, vulnerability.vulnerabilitydescription)
-        vulnerability.POC = get_subdoc(doc, vulnerability.POC)
-        vulnerability.vulnerabilitysolution = get_subdoc(doc, vulnerability.vulnerabilitysolution)
-        vulnerability.vulnerabilityreferlnk = get_subdoc(doc, vulnerability.vulnerabilityreferlnk)
+        vulnerability.vulnerabilitydescription = get_subdoc(doc, vulnerability.vulnerabilitydescription,token, base_url)
+        vulnerability.POC = get_subdoc(doc, vulnerability.POC, token,base_url)
+        vulnerability.vulnerabilitysolution = get_subdoc(doc, vulnerability.vulnerabilitysolution, token,base_url)
+        vulnerability.vulnerabilityreferlnk = get_subdoc(doc, vulnerability.vulnerabilityreferlnk,token, base_url)
 
         vulnerability.instances_data = [
             {
@@ -164,7 +92,7 @@ def generate_vulnerability_document(pk,Report_type,standard):
             for instance in Vulnerableinstance.objects.filter(vulnerabilityid=vulnerability, project=project)
             if instance.URL  # Exclude instances with empty URL
         ]
-        currentdate=datetime.now()
+        
     totalretest = [
     {
         "startdate": retest.startdate,
@@ -172,7 +100,8 @@ def generate_vulnerability_document(pk,Report_type,standard):
         "owners": [owner.full_name for owner in retest.owner.all()]
     }
     for retest in totalretests_queryset
-]
+    ]
+    currentdate=datetime.now()
     context = {'project': project, 'vulnerabilities': vuln,'Report_type':Report_type,'mycomany':mycomany,'projectmanagers':projectmanagers,'customeruser':customeruser,'owners': owners,
               'project_exception':project_exception,'project_description':project_description,"settings":settings,"currentdate":currentdate,'value2':'10',
                'standard':standard,'totalvulnerability':totalvulnerability,'totalretest':totalretest,'projectscope':projectscope,
@@ -182,28 +111,130 @@ def generate_vulnerability_document(pk,Report_type,standard):
     jinja_env.trim_blocks = True
     jinja_env.lstrip_blocks = True
     doc.render(context,jinja_env)
-    font = doc.styles['List Bullet'].font
-    font.name = 'Calibri'
-    font.size = Pt(16)
-    section = doc.sections[1]
-    #section.left_margin = Inches(1)
-    #section.right_margin = Inches(1)
-    section.top_margin = Inches(1)
-    section.bottom_margin = Inches(1)
-    for table in doc.tables:
-        for row in table.rows:
-            if all(cell.text.strip() == "" for cell in row.cells):
-                row_element = row._element
-                row_element.getparent().remove(row_element)
-            else:
-                # Set row height to fit content
-                row.height = None  # Automatic height
-                row.height_rule = None  # Automatically adjust height
+    doc = main_doc_style(doc)
     response = HttpResponse(content_type='application/vnd.openxmlformats-officedocument.wordprocessingml.document')
     response['Content-Disposition'] = f'attachment; filename={project.name}vulnerability_report.docx'
     doc.save(response)
     return response
 
+
+def CreateExcel(pk):
+    project_name = Project.objects.values_list('name', flat=True).get(id=pk)
+    instances = Vulnerableinstance.objects.filter(project_id=pk).select_related('vulnerabilityid').order_by('-vulnerabilityid__cvssscore')
+    output = io.BytesIO()
+    book = Workbook(output)
+    sheet = book.add_worksheet('Vulnerabilities')
+    wrap_format = book.add_format({'text_wrap': True, 'align': 'center', 'valign': 'vcenter'})
+    sheet.set_column('A:A', 20)
+    sheet.set_column('C:C', 15)
+    sheet.set_column('D:D', 15)
+    sheet.set_column('E:E', 15)
+    sheet.set_column('G:G', 70)
+    sheet.set_column('H:H', 70)
+
+    # Write header
+    header = ['Title', 'Severity', 'Status', 'URL/IP', 'Parameter/Port', 'CVSS Score', 'Description','Recommendation']
+    for col_num, col_value in enumerate(header):
+        sheet.write(0, col_num, col_value,wrap_format)
+
+    # Write data rows
+    for row_num, instance in enumerate(instances, start=1):
+        row_data = [
+            instance.vulnerabilityid.vulnerabilityname,
+            instance.vulnerabilityid.vulnerabilityseverity,
+            instance.status,
+            instance.URL,
+            instance.Parameter,
+            instance.vulnerabilityid.cvssscore,
+            html.unescape(bleach.clean(instance.vulnerabilityid.vulnerabilitydescription, tags=[], strip=True)),
+            html.unescape(bleach.clean(instance.vulnerabilityid.vulnerabilitysolution, tags=[], strip=True))
+        ]
+        for col_num, col_value in enumerate(row_data):
+            sheet.write(row_num, col_num, col_value, wrap_format)
+
+
+    book.close()
+    response = HttpResponse(output.getvalue(), content_type='application/vnd.openxmlformats-officedocument.spreadsheetml.sheet')
+    response['Content-Disposition'] = f'attachment; filename={project_name}vulnerability_report.xlsx'
+
+    return response
+
+
+
+
+def GetHTML(Report_type,pk,standard,request):
+    project = Project.objects.get(pk=pk)
+
+    ## Get All Projects Vulnerability filter higher to lower CVSS Score
+    vuln = Vulnerability.objects.filter(project=project).order_by('-cvssscore')
+
+    #### Get ALL Instances for the Projects  (Not using Vulnerability colum)  ### Need to optimize to speed up the HTML generation
+    instances = Vulnerableinstance.objects.filter(project=project)
+    project_manager_group = CustomGroup.objects.get(name='Project Manager')
+
+    projectmanagers = CustomUser.objects.filter(groups=project_manager_group)
+    customeruser =  CustomUser.objects.filter(company=project.companyname, is_active=True)
+
+
+    ciritcal =  vuln.filter(project=project,vulnerabilityseverity='Critical',status='Vulnerable').count()
+    high =  vuln.filter(project=project,vulnerabilityseverity='High',status='Vulnerable').count()
+    medium =  vuln.filter(project=project,vulnerabilityseverity='Medium',status='Vulnerable').count()
+    low =  vuln.filter(project=project,vulnerabilityseverity='Low',status='Vulnerable').count()
+    info = vuln.filter((Q(status='Vulnerable')) & (Q(vulnerabilityseverity='Informational') | Q(vulnerabilityseverity='None'))).count()
+
+    custom_style = Style(
+    colors=("#FF491C", "#F66E09", "#FBBC02", "#20B803", "#3399FF"),
+    background='transparent',
+    plot_background='transparent',
+    legend_font_size=0,
+    legend_box_size=0,
+    value_font_size=40
+    )
+    pie_chart = pygal.Pie(style=custom_style)
+    pie_chart.legend_box_size = 0
+
+    pie_chart.add('Critical', ciritcal)
+    pie_chart.add('High', high)
+    pie_chart.add('Medium', medium)
+    pie_chart.add('Low', low)
+    pie_chart.add('Informational', info)
+
+
+    ### Get Total Vulnerability Count
+    totalvulnerability = vuln.filter(project=project).count()
+    mycomany = Company.objects.filter(internal=True).values_list('name', flat=True).first()
+
+
+    ### Get All Scope from the project
+    projectscope = PrjectScope.objects.filter(project=project)
+
+    ## Get Retest Details
+    totalretest = ProjectRetest.objects.filter(project_id=pk)
+    data = {'projectscope':projectscope,'totalvulnerability':totalvulnerability,'standard':standard,'Report_type':Report_type,'mycomany':mycomany,
+            'totalretest':totalretest,'vuln':vuln,'project':project,"settings":settings,'ciritcal':ciritcal,'high':high,
+            'medium':medium,'low':low,'info':info,'instances':instances,'projectmanagers':projectmanagers,'customeruser':customeruser,'pie_chart':pie_chart.render(is_unicode=True)}
+    try:
+        # Render the template to a string
+        rendered_content = render_to_string('report.html', data, request=request)
+        response = generate_pdf_report(rendered_content,base_url)
+        response['Content-Disposition'] = f'attachment; filename={project.name}vulnerability_report.pdf'
+        return response
+    except (TemplateDoesNotExist, IOError) as e:
+        # Handle template not found error
+        logger.error("Something Went Wrong: %s", e)
+        return HttpResponse("Something Went Wrong", status=500)
+
+
+def generate_pdf_report(rendered_content,base_url):
+    try:
+        pdf = HTML(string=rendered_content,url_fetcher=my_fetcher,base_url=base_url).write_pdf()
+        # Return the PDF response
+        response = HttpResponse(pdf, content_type='application/pdf')
+        return response
+    except Exception as e:
+        # Return a server error response if there's an issue
+        logger.error("Something Went Wrong: %s", e)
+        return HttpResponse("Something went wrong", status=500)
 
 
 def is_whitelisted(url):
@@ -260,161 +291,3 @@ def my_fetcher(url):
             }
     else:
         raise ValueError(f'URL is Not WhiteListed for: {url!r}')
-
-
-# Global variable to store base_url
-base_url = ""
-
-def get_base_url(request=None):
-    """Generate and return the base URL based on settings and request."""
-    global base_url
-    if settings.USE_DOCKER:
-        base_url = "https://nginx/"
-    else:
-        if request:
-            base_url = f"{request.scheme}://{request.get_host()}"
-        else:
-            raise Exception("Request is required when not using Docker.")
-    return base_url
-
-
-
-def CheckReport(Report_format,Report_type,pk,url,standard,request):
-    get_base_url()
-    if Report_format == "excel":
-        response =  CreateExcel(pk)
-    if Report_format == "docx":
-        response = generate_vulnerability_document(pk,Report_type,standard)
-    if Report_format == "pdf":
-        global token
-        auth_header = request.headers.get('Authorization')
-    
-        if auth_header:
-        # Extract token from "Bearer <token>"
-            token = auth_header.split(' ')[1] if auth_header.startswith('Bearer ') else None
-        response = GetHTML(Report_format,Report_type,pk,url,standard,request)
-    
-    return response
-
-
-def CreateExcel(pk):
-    instances = Vulnerableinstance.objects.filter(project_id=pk).select_related('vulnerabilityid').order_by('-vulnerabilityid__cvssscore')
-    output = io.BytesIO()
-    book = Workbook(output)
-    sheet = book.add_worksheet('Vulnerabilities')
-    wrap_format = book.add_format({'text_wrap': True, 'align': 'center', 'valign': 'vcenter'})
-    sheet.set_column('A:A', 20)
-    sheet.set_column('C:C', 15)
-    sheet.set_column('D:D', 15)
-    sheet.set_column('E:E', 15)
-    sheet.set_column('G:G', 70)
-    sheet.set_column('H:H', 70)
-
-    # Write header
-    header = ['Title', 'Severity', 'Status', 'URL/IP', 'Parameter/Port', 'CVSS Score', 'Description','Recommendation']
-    for col_num, col_value in enumerate(header):
-        sheet.write(0, col_num, col_value,wrap_format)
-
-    # Write data rows
-    for row_num, instance in enumerate(instances, start=1):
-        row_data = [
-            instance.vulnerabilityid.vulnerabilityname,
-            instance.vulnerabilityid.vulnerabilityseverity,
-            instance.status,
-            instance.URL,
-            instance.Parameter,
-            instance.vulnerabilityid.cvssscore,
-            html.unescape(bleach.clean(instance.vulnerabilityid.vulnerabilitydescription, tags=[], strip=True)),
-            html.unescape(bleach.clean(instance.vulnerabilityid.vulnerabilitysolution, tags=[], strip=True))
-        ]
-        for col_num, col_value in enumerate(row_data):
-            sheet.write(row_num, col_num, col_value, wrap_format)
-
-
-    book.close()
-    response = HttpResponse(output.getvalue(), content_type='application/vnd.openxmlformats-officedocument.spreadsheetml.sheet')
-    response['Content-Disposition'] = 'attachment; filename=instances.xlsx'
-
-    return response
-
-
-
-
-def GetHTML(Report_format,Report_type,pk,url,standard,request):
-    project = Project.objects.get(pk=pk)
-
-    ## Get All Projects Vulnerability filter higher to lower CVSS Score
-    vuln = Vulnerability.objects.filter(project=project).order_by('-cvssscore')
-
-    #### Get ALL Instances for the Projects  (Not using Vulnerability colum)  ### Need to optimize to speed up the HTML generation
-    instances = Vulnerableinstance.objects.filter(project=project)
-    project_manager_group = CustomGroup.objects.get(name='Project Manager')
-
-    projectmanagers = CustomUser.objects.filter(groups=project_manager_group)
-    customeruser =  CustomUser.objects.filter(company=project.companyname, is_active=True)
-
-
-    ciritcal =  vuln.filter(project=project,vulnerabilityseverity='Critical',status='Vulnerable').count()
-    high =  vuln.filter(project=project,vulnerabilityseverity='High',status='Vulnerable').count()
-    medium =  vuln.filter(project=project,vulnerabilityseverity='Medium',status='Vulnerable').count()
-    low =  vuln.filter(project=project,vulnerabilityseverity='Low',status='Vulnerable').count()
-    info = vuln.filter((Q(status='Vulnerable')) & (Q(vulnerabilityseverity='Informational') | Q(vulnerabilityseverity='None'))).count()
-
-    custom_style = Style(
-    colors=("#FF491C", "#F66E09", "#FBBC02", "#20B803", "#3399FF"),
-    background='transparent',
-    plot_background='transparent',
-    legend_font_size=0,
-    legend_box_size=0,
-    value_font_size=40
-    )
-    pie_chart = pygal.Pie(style=custom_style)
-    pie_chart.legend_box_size = 0
-
-    pie_chart.add('Critical', ciritcal)
-    pie_chart.add('High', high)
-    pie_chart.add('Medium', medium)
-    pie_chart.add('Low', low)
-    pie_chart.add('Informational', info)
-
-
-    ### Get Total Vulnerability Count
-    totalvulnerability = vuln.filter(project=project).count()
-    mycomany = Company.objects.filter(internal=True).values_list('name', flat=True).first()
-
-
-    ### Get All Scope from the project
-    projectscope = PrjectScope.objects.filter(project=project)
-
-    ## Get Retest Details
-    totalretest = ProjectRetest.objects.filter(project_id=pk)
-    data = {'projectscope':projectscope,'totalvulnerability':totalvulnerability,'standard':standard,'Report_type':Report_type,'mycomany':mycomany,
-            'totalretest':totalretest,'vuln':vuln,'project':project,"settings":settings,"url":url,'ciritcal':ciritcal,'high':high,
-            'medium':medium,'low':low,'info':info,'instances':instances,'projectmanagers':projectmanagers,'customeruser':customeruser,'pie_chart':pie_chart.render(is_unicode=True)}
-    if settings.USE_DOCKER:
-        base_url = "https://nginx/"
-    else:
-        base_url = f"{request.scheme}://{request.get_host()}"
-    
-    try:
-        # Render the template to a string
-        rendered_content = render_to_string('report.html', data, request=request)
-        response = generate_pdf_report(rendered_content,base_url)
-        return response
-    except (TemplateDoesNotExist, IOError) as e:
-        # Handle template not found error
-        logger.error("Something Went Wrong: %s", e)
-        return HttpResponse("Something Went Wrong", status=500)
-
-
-def generate_pdf_report(rendered_content,base_url):
-    try:
-        pdf = HTML(string=rendered_content,url_fetcher=my_fetcher,base_url=base_url).write_pdf()
-        # Return the PDF response
-        response = HttpResponse(pdf, content_type='application/pdf')
-        response['Content-Disposition'] = 'attachment; filename="report.pdf"'
-        return response
-    except Exception as e:
-        # Return a server error response if there's an issue
-        logger.error("Something Went Wrong: %s", e)
-        return HttpResponse("Something went wrong", status=500)
